+++ pdfimpose/errors.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright Louis Paternault 2011-2015, 2017
+# Copyright Louis Paternault 2011-2017
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by