+++ CHANGELOG.md
@@ -5,6 +5,8 @@
     * No longer crash when using pdfimpose on file without any metadata (closes #12).
     * Warn user if all pages do not have the same dimension (closes #11).
     * Display nicer messages with several input-file related errors (absent, unreadable, malformed, etc. file).
+    * Add options `--paper` and `--sheets`, to define how document is folded more easily (closes #7).
+    * Horizontal and vertical folds are alternated as much as possible (closes #8).
 
     -- Louis Paternault <spalax+python@gresille.org>